+++ src/main/java/nextstep/ladder/view/OutputView.java
@@ -1,6 +1,7 @@
 package nextstep.ladder.view;
 
 import nextstep.ladder.domain.ladder.Ladder;
+import nextstep.ladder.domain.ladder.Line;
 import nextstep.ladder.domain.participant.Participants;
 
 public class OutputView {
@@ -8,11 +9,28 @@ public class OutputView {
 	public void printResult(Participants participants, Ladder ladder) {
 		System.out.println("실행결과");
 
-		participants.forEach(participant -> System.out.print(participant + "  "));
+		participants.getNames()
+			.forEach(participant -> System.out.print(participant + "  "));
 		System.out.println();
 
-		ladder.forEach(line -> {
-			System.out.println(line.getLine());
-		});
+		ladder.getLadder()
+				.forEach(line -> System.out.println(getLine(line)));
+	}
+
+	public StringBuilder getLine(Line line) {
+		StringBuilder stringBuilder = new StringBuilder();
+		stringBuilder.append("|");
+		line.getPoints().forEach(point -> appendLine(point, stringBuilder));
+		return stringBuilder;
+	}
+
+	private void appendLine(Boolean point, StringBuilder stringBuilder) {
+		if (point){
+			stringBuilder.append("------");
+		}
+		else{
+			stringBuilder.append("      ");
+		}
+		stringBuilder.append("|");
 	}
 }
